Add .gitignore & remove unnecessary files
This commit adds the default .gitignore settings for a Python project.
It also deletes cache files like __pycache__ and db.sqlite3 that were
polluting the file tree. These will automatically be generated when run
on the user's computer (migrations may need to be regenerated manually,
recommend including that in setup instructions in README.md)